Preparation for GUI Refresh
Added a refresh function to set changes to disk for the database.
Moved GUI form setup into the GUI files.
